+++ jugger/lib/jugger.dart
@@ -1,6 +1,15 @@
 library jugger;
 
 export 'src/annotations.dart'
-    show module, provide, singleton, inject, Component, Inject, bind, componentBuilder, Named;
+    show
+        module,
+        provide,
+        singleton,
+        inject,
+        Component,
+        Inject,
+        bind,
+        componentBuilder,
+        Named;
 
 export 'src/provider.dart' show Provider, SingletonProvider, IProvider;