+++ jugger/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0-alpha
+
+* Add annotations: ComponentBuilder, Named
+
 ## 1.0.0-alpha
 
 * First release